# nodeskclaw-knowledge/contracts/frontend/v1.0.0/openapi.frontend.yaml
openapi: 3.1.0
info:
  title: NodeSKClaw Knowledge Frontend Contract
  version: 1.0.0
  description: Curated frontend-facing API surface. Contract version is independent from HTTP /api/v1 and /api/v2 path versions.
servers:
  -
    url: "{knowledgeBaseUrl}"
    variables:
      knowledgeBaseUrl:
        default: "http://127.0.0.1:4530"
security:
  -
    bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ErrorResponse:
      title: ErrorResponse
      type: object
      properties:
        code:
          type: integer
          minimum: 40000
        error_code:
          type: integer
          minimum: 40000
        message_key:
          type: string
        message:
          type: string
        data:
          type: "null"
        details:
          type: object
        message_params:
          type: object
          additionalProperties:
            type: string
      additionalProperties: false
      required:
        - code
        - error_code
        - message_key
        - message
        - data
    PageData:
      title: PageData
      type: object
      properties:
        items:
          type: array
        total:
          type: integer
          minimum: 0
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
      additionalProperties: false
      required:
        - items
        - total
        - page
        - page_size
...
    KnowledgeBaseUpdate:
      title: KnowledgeBaseUpdate
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
        description:
          type:
            - string
            - "null"
        tags:
          type:
            - array
            - "null"
          items:
            type: string
        visibility:
          type: string
          enum:
            - private
            - department
            - organization
      additionalProperties: false
...
paths:
...
  /api/v2/knowledge-bases:
    get:
      operationId: KB01
      summary: GET /api/v2/knowledge-bases
      tags:
        - knowledge-base
      x-stability: stable
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                properties:
                  code:
                    type: integer
                  error_code:
                    type:
                      - integer
                      - "null"
                  message_key:
                    type:
                      - string
                      - "null"
                  message:
                    type: string
                  data:
                    allOf:
                      -
                        $ref: "#/components/schemas/PageData"
                      -
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: "#/components/schemas/KnowledgeBase"
                  details:
                    type: object
                  message_params:
                    type: object
                additionalProperties: false
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        503:
          description: Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: KB02
      summary: POST /api/v2/knowledge-bases
      tags:
        - knowledge-base
      x-stability: stable
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeBaseCreate"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                properties:
                  code:
                    type: integer
                  error_code:
                    type:
                      - integer
                      - "null"
                  message_key:
                    type:
                      - string
                      - "null"
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/KnowledgeBase"
                  details:
                    type: object
                  message_params:
                    type: object
                additionalProperties: false
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        503:
          description: Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  "/api/v2/knowledge-bases/{kb_id}":
    get:
      operationId: KB03
      summary: "GET /api/v2/knowledge-bases/{kb_id}"
      tags:
        - knowledge-base
      x-stability: stable
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                properties:
                  code:
                    type: integer
                  error_code:
                    type:
                      - integer
                      - "null"
                  message_key:
                    type:
                      - string
                      - "null"
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/KnowledgeBase"
                  details:
                    type: object
                  message_params:
                    type: object
                additionalProperties: false
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        503:
          description: Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: KB04
      summary: "PATCH /api/v2/knowledge-bases/{kb_id}"
      tags:
        - knowledge-base
      x-stability: stable
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeBaseUpdate"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                properties:
                  code:
                    type: integer
                  error_code:
                    type:
                      - integer
                      - "null"
                  message_key:
                    type:
                      - string
                      - "null"
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/KnowledgeBase"
                  details:
                    type: object
                  message_params:
                    type: object
                additionalProperties: false
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        503:
          description: Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  "/api/v1/knowledge-bases/{kb_id}":
    delete:
      operationId: KB05
      summary: "DELETE /api/v1/knowledge-bases/{kb_id}"
      tags:
        - knowledge-base
      x-stability: compatibility
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                properties:
                  code:
                    type: integer
                  error_code:
                    type:
                      - integer
                      - "null"
                  message_key:
                    type:
                      - string
                      - "null"
                  message:
                    type: string
                  data:
                    type: "null"
                  details:
                    type: object
                  message_params:
                    type: object
                additionalProperties: false
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        503:
          description: Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  "/api/v1/knowledge-bases/{kb_id}/acl":
    get:
      operationId: KB06
      summary: "GET /api/v1/knowledge-bases/{kb_id}/acl"
      tags:
        - acl
      x-stability: compatibility
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                properties:
                  code:
                    type: integer
                  error_code:
                    type:
                      - integer
                      - "null"
                  message_key:
                    type:
                      - string
                      - "null"
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Acl"
                  details:
                    type: object
                  message_params:
                    type: object
                additionalProperties: false
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        503:
          description: Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: KB07
      summary: "POST /api/v1/knowledge-bases/{kb_id}/acl"
      tags:
        - acl
      x-stability: compatibility
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AclCreate"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                properties:
                  code:
                    type: integer
                  error_code:
                    type:
                      - integer
                      - "null"
                  message_key:
                    type:
                      - string
                      - "null"
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/Acl"
                  details:
                    type: object
                  message_params:
                    type: object
                additionalProperties: false
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        503:
          description: Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  "/api/v1/knowledge-bases/{kb_id}/acl/{acl_id}":
    delete:
      operationId: KB08
      summary: "DELETE /api/v1/knowledge-bases/{kb_id}/acl/{acl_id}"
      tags:
        - acl
      x-stability: compatibility
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - data
                properties:
                  code:
                    type: integer
                  error_code:
                    type:
                      - integer
                      - "null"
                  message_key:
                    type:
                      - string
                      - "null"
                  message:
                    type: string
                  data:
                    type: "null"
                  details:
                    type: object
                  message_params:
                    type: object
                additionalProperties: false
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        503:
          description: Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  "/api/v1/source-files/{source_file_id}/download":
    get:
      operationId: F10
      summary: "GET /api/v1/source-files/{source_file_id}/download"
      tags:
        - documents
      x-stability: compatibility
      responses:
        200:
          description: Success
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        503:
          description: Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...